requirements.txt
numpy==1.25.0
pandas==2.0.3
Pillow==9.5.0
streamlit==1.24.0
